separated MarcRecord from other models
